pyproject.toml
[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.dev-dependencies]
dev = [
    "watchdog",
    "faker",
    "flask-debugtoolbar>=0.13.1",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "lijialiang", email = "614237720@qq.com"},
]
dependencies = [
    "click==8.1.3",
    "coverage==6.4.2",
    "flask==2.1.3",
    "flask-login==0.6.1",
    "flask-sqlalchemy==2.5.1",
    "importlib-metadata==4.12.0",
    "itsdangerous>=2.1.2",
    "jinja2==3.1.2",
    "markupsafe==2.1.1",
    "python-dotenv==0.20.0",
    "sqlalchemy==1.4.39",
    "werkzeug==2.1.2",
    "zipp==3.8.1",
    "flask-ckeditor>=0.4.6",
    "wtforms>=3.0.1",
    "flask-migrate>=4.0.4",
    "flask-mail>=0.9.1",
    "flask-wtf>=1.1.1",
    "email-validator>=1.3.1",
    "web3>=6.0.0",
    "unidecode>=1.3.6",
    "pillow>=9.4.0",
    "flask-avatars>=0.2.2",
    "flask-dropzone>=1.6.0",
    "bootstrap-flask>=2.2.0",
    "flask-moment>=1.0.5",
    "flask-whooshee>=0.8.2",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
